+++ index.js
@@ -1,7 +1,9 @@
-var express     = require ("express"),
-    app         = express (),
-    bodyParser  = require ("body-parser"),
-    mongoose    = require ("mongoose"),
+var express          = require ("express"),
+    app              = express (),
+    bodyParser       = require ("body-parser"),
+    mongoose         = require ("mongoose"),
+    methodOverride   = require ("method-override"),
+    expressSanitizer = require ("express-sanitizer"),
 
     // authentication
     passport        = require ("passport"),
@@ -24,6 +26,8 @@ app.set ("view engine", "ejs");
 
 app.use (express.static(__dirname + "/public"));
 app.use (bodyParser.urlencoded( { extended : true } ));
+app.use (methodOverride("_method"));
+app.use (expressSanitizer());
 
 // authentication
 app.use (require("express-session")({