+++ models/comment.js
@@ -2,12 +2,10 @@ var mongoose = require ("mongoose");
 
 var commentSchema = new mongoose.Schema({
     text: String,
-    author: [
-        {
-            type: mongoose.Schema.Types.ObjectId,
-            ref: "User"
-        }
-    ]
+    author: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: "User"
+    }
 });
 
 module.exports = mongoose.model("Comment", commentSchema);