+++ routes/comments.js
@@ -28,6 +28,8 @@ router.post ("/", middleware.isLoggedIn, function (req, res) {
         if (error) {
             console.log ("Couldn't find corresponding title " + error);
         } else {
+            req.body.comment.text = req.sanitize(req.body.comment.text);
+            
             Comment.create (req.body.comment, function (error, comment) {
                 if (error) {
                     console.log ("Couldn't create comment " + error);
@@ -52,5 +54,54 @@ router.post ("/", middleware.isLoggedIn, function (req, res) {
         }
     });    
 });
+
+// comments EDIT
+router.get ("/:idComment/edit", middleware.isCommentOwner, function (req, res) {
+    Comment.findById (req.params.idComment).populate("author").exec (function (error, comment) {
+        if (error) {
+            console.log ("Couldn't find comment to edit " + error);
+        } else {
+            res.render ("comments/edit", {comment : comment, title : req.params.id});
+        }
+    });
+});
+
+// comments UPDATE
+router.put ("/:idComment", middleware.isCommentOwner, function (req, res) {
+    
+    req.body.comment.text = req.sanitize(req.body.comment.text);
+    
+    Comment.findByIdAndUpdate (req.params.idComment, req.body.comment, function (error, comment) {
+        if (error) {
+            console.log ("Couldn't find and update comment " + error);
+        } else {
+            res.redirect ("/movies/" + req.params.id);
+        }
+    });
+});
+
+
+// comments DESTROY
+router.delete ("/:idComment", middleware.isCommentOwner, function (req, res) {
+    Comment.findByIdAndRemove (req.params.idComment, function (error, comment) {
+        if (error) {
+            console.log ("Couldn't find and delete comment " + error);
+        } else {
+            // also deleting reference from title
+            Tvtitle.update (
+                { "comments" : req.params.idComment },
+                { "$pull" : { "comments" : req.params.idComment } },
+                function (error, title) {
+                    if (error) {
+                        console.log ("Couldn't update title to remove deleted comment");
+                    } else {
+                        res.redirect("/movies/" + req.params.id);
+                    }
+                }
+            );
+        }
+    });
+});
+
 
 module.exports = router;