+++ routes/movies.js
@@ -15,7 +15,7 @@ router.get ("/", function (req, res) {
     //get all movies from the database
     Tvtitle.find ({}, function (error, movies) {
         if (error) {
-            console.log (error);
+            console.log ("Couldn't find all " + error);
         } else {
             res.render("movies/index", {movies : movies});
         }
@@ -29,20 +29,13 @@ router.get ("/new", middleware.isLoggedIn, function (req, res) {
 
 // CREATE route - Add new movie to db
 router.post ("/", middleware.isLoggedIn, function (req, res) {
-    // get data from form POST action
-    var title = new Tvtitle ({
-        name: req.body.name,
-        image: req.body.image,
-        slug: slugify(req.body.name),
-        description: req.body.description,
-        author: req.user._id
-    });
+    var title = req.body.title;
+    title.slug = slugify (req.body.title.name);
+    title.author = req.user._id;
     
-    // save to database
-    title.save ( function (error, title) {
+    Tvtitle.create (title, function (error, title) {
         if (error) {
-            console.log("Something went wrong");
-            console.log(error);
+            console.log("Couldn't create title " + title + error);
         } else {
             res.redirect ("/movies");
         }
@@ -65,6 +58,60 @@ router.get ("/:id", function (req, res) {
                 // render show page:
                 res.render ("movies/show", {title : title});                
             });
+        }
+    });
+});
+
+// EDIT route - show form to edit
+router.get ("/:id/edit", middleware.isOwner, function (req, res) {
+    Tvtitle.findById (req.params.id).populate("author").exec (function (error, title) {
+        if (error) {
+            console.log ("Couldn't find title to edit " + error);
+        } else {
+            res.render ("movies/edit", {title : title});
+        }
+    });
+});
+
+// UPDATE route - update title
+router.put ("/:id", middleware.isOwner, function (req, res) {
+    //re-slugify for updated title!
+    req.body.title.slug = slugify(req.body.title.name);
+    
+    Tvtitle.findByIdAndUpdate (req.params.id, req.body.title, function (error, title) {
+        if (error) {
+            console.log ("Couldn't find and update " + error);
+        } else {
+            res.redirect ("/movies/" + req.params.id);
+        }
+    });
+});
+
+// DESTROY route
+router.delete ("/:id", middleware.isOwner, function (req, res) {
+    
+    // delete all associated comments
+    Tvtitle.findById (req.params.id).populate("comments").exec (function (error, title) {
+        if (error) {
+            console.log ("Couldn't show title page " + error);
+        } else {
+            title.comments.forEach (function (comment) {
+                Comment.findByIdAndRemove(comment._id, function (error) {
+                    if (error) {
+                        console.log ("Couldn't delete comment " + comment._id + error);
+                    } else {
+                        console.log ("Deleted " + comment._id);
+                    }
+                });
+            });
+        }
+    });
+    
+    Tvtitle.findByIdAndRemove(req.params.id, function (error, title) {
+        if (error) {
+            console.log ("Couldn't delete title " + error);
+        } else {
+            res.redirect ("/movies");
         }
     });
 });